cagov/odi-engineering#65 Organize config and documentation. Delete references to wordpress-to-github config.
