+++ .vscode/settings.json
@@ -1,0 +1,3 @@
+{
+  "workbench.colorTheme": "Quiet Light"
+}
