+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "odi-publishing-test",
-  "version": "1.0.0",
+  "version": "0.0.1",
   "description": "",
   "main": "index.js",
   "type": "module",
